Ray marching performance update
The main change is in the "ray has hit the surface" condition, I modified it, so the marching stops when the ray is closer then the visible distance per pixel from the camera at this position. Since we march far less steps it gives quite a considerable performance boost. In my case it was about 20-100% more fps, depending on the distance from the fractal. It didn't work by itself though, since it gave some weird graphical artifacts, the problem was in the fact that the point was too far from the fractal surface, so I added an additional step to march in the opposite direction of the surface normal to get to the closest surface point. (Actually, isn't it simpler to find the collisions between the marble and the fractal in this way?)  
Also I added a faster numerical normal function, it uses 4 DE calculations instead of 6, credit to Inigo Quilez.